requirements.txt
Django==1.11.6
django-cors-headers==2.4.1
django-restframework==0.0.1
djangorestframework==3.10.3
pytz==2019.3
